+++ R/box_unused_attached_mod_linter.R
@@ -82,6 +82,9 @@ box_unused_attached_mod_linter <- function() {
     attached_modules <- get_attached_modules(xml)
     attached_three_dots <- get_attached_mod_three_dots(xml)
     function_calls <- get_function_calls(xml)
+    glue_object_calls <- get_objects_in_strings(xml)
+
+    all_calls_text <- c(function_calls$text, glue_object_calls)
 
     unused_module <- lapply(attached_modules$xml, function(attached_module) {
       module_text <- basename(lintr::get_r_string(attached_module))
@@ -93,7 +96,7 @@ box_unused_attached_mod_linter <- function() {
         sep = "$"
       )
 
-      functions_used <- length(intersect(func_list, function_calls$text))
+      functions_used <- length(intersect(func_list, all_calls_text))
 
       if (functions_used == 0) {
         lintr::xml_nodes_to_lints(
@@ -109,7 +112,7 @@ box_unused_attached_mod_linter <- function() {
       module_text <- basename(lintr::get_r_string(attached_module))
       module_text <- sub("\\[\\.\\.\\.\\]", "", module_text)
       func_list <- attached_three_dots$nested[[module_text]]
-      functions_used <- length(intersect(func_list, function_calls$text))
+      functions_used <- length(intersect(func_list, all_calls_text))
 
       if (functions_used == 0) {
         lintr::xml_nodes_to_lints(