+++ NEWS.md
@@ -1,5 +1,6 @@
 # box.linters (development version)
 
+* Added handler for `glue` string templates.
 * Added box_mod_fun_exists_linter() to default linters 
 * [bug fix] box_trailing_commas_linter() now properly handles a #nolint for other linters
 * [bug fix] `box_unused_att_pkg_fun_linter()` had issues with lists of functions. Linter function